bot/requirements.txt
colorama
discord.py
ruamel.yaml
